Fixed bug when making credit note
The module passes two objects as argument to _compute_deliver_status and def _compute_shipment_status when client is creating a credit note. I tried to fix the bug by adding self = self[1], as it appears the right instance is always at index[1]. To avoid error when only one object is passed.